backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
psycopg2-binary>=2.9
django-jazzmin>=3.0
